Jenkins-server automated installation script
Run this script after connecting to the freshly created AWS (or any cloud provider) server. It will automatically install NGINX, set that up as a proxy for Jenkins, install Jenkins and it will give you the default admin password to connect to your Jenkins server. You can change that after the first connection.