Challenge 2
You're on a team developing an algorithm that calculates the value of pi. The code for the project is being stored on a GitHub repository. The team wants to test the latest changes to the algorithm with every push to the repo. They also want to display the status of that most recent test directly in the repo's README file